added version to rosetta operation
